# public/knowledge-api-openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge Management API
  description: |
    RESTful API for the knowledge management system.
    
    ## Overview
    This API provides endpoints for:
    - **Search** for full-text and semantic search across knowledge articles
    - **Articles** for article CRUD operations
    - **Analytics** - for search queries, results, agent interactions, and usage metrics
    
    ## Authentication
    All endpoints require authentication via JWT Bearer tokens obtained through OAuth 2.0 / OpenID Connect.
    
    ## Rate Limiting
    - Standard tier: 100 requests/minute per user
    - Professional tier: 1,000 requests/minute per user
    - Enterprise tier: Custom limits
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets
    
    ## Versioning
    API versioning is done via URL path (`/api/v1/`). Breaking changes will increment the major version.
    Deprecation notices are provided via `Sunset` header 6 months before removal.

  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
...
security:
  - bearerAuth: []
paths:
...
  /articles:
    get:
      tags:
        - Articles
      summary: List all articles
      description: |
        Retrieves a paginated list of articles. Supports filtering and sorting.
        
        **Authorization:** Requires `articles:read` scope.
      operationId: listArticles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/CategoryParam'
        - name: status
          in: query
          description: Filter by publication status
          schema:
            type: string
            enum: [draft, published, archived]
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [createdDate, lastUpdated, title, viewCount]
            default: lastUpdated
        - name: sortOrder
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of articles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Articles
      summary: Create a new article
      description: |
        Creates a new knowledge article.
        
        **Authorization:** Requires `articles:write` scope.
        
        **Workflow:**
        1. Article created in `draft` status
        2. Content validated (length, format, required fields)
        3. AI generates embeddings asynchronously
        4. Article available for search after publishing
      operationId: createArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCreateRequest'
            example:
              title: "How to Handle International Shipping"
              content: "This guide covers international shipping procedures..."
              category: "Shipping"
              tags: ["international", "shipping", "customs"]
              status: "draft"
      responses:
        '201':
          description: Article created
          headers:
            Location:
              description: URL of the created article
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /articles/{articleId}:
    parameters:
      - $ref: '#/components/parameters/ArticleIdParam'

    get:
      tags:
        - Articles
      summary: Get article by ID
      description: |
        Retrieves a single article by its ID.
        
        **Authorization:** Requires `articles:read` scope.
        
        **Side Effects:**
        - Increments view count (debounced per user, 5-minute window)
        - Logs view event for analytics
      operationId: getArticle
      parameters:
        - name: includeHistory
          in: query
          description: Include version history
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Article details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Articles
      summary: Update an article
      description: |
        Updates an existing article. Creates a new version in history.
        
        **Authorization:** Requires `articles:write` scope.
        
        **Versioning:**
        - Each update creates a new version
        - Previous versions retained for 90 days
        - Version history accessible via `includeHistory` parameter
        
        **AI Processing:**
        - Embeddings regenerated asynchronously on content change
        - Search index updated within 30 seconds
      operationId: updateArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleUpdateRequest'
      responses:
        '200':
          description: Article updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - Articles
      summary: Delete an article
      description: |
        Soft-deletes an article (moves to archived status).
        
        **Authorization:** Requires `articles:delete` scope.
        
        **Behavior:**
        - Article removed from search index immediately
        - Data retained for 30 days before permanent deletion
        - Can be restored via support request within retention period
      operationId: deleteArticle
      responses:
        '204':
          description: Article deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /articles/{articleId}/publish:
    post:
      tags:
        - Articles
      summary: Publish an article
      description: |
        Publishes a draft article, making it searchable.
        
        **Authorization:** Requires `articles:publish` scope.
      operationId: publishArticle
      parameters:
        - $ref: '#/components/parameters/ArticleIdParam'
      responses:
        '200':
          description: Article published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          description: Article cannot be published (e.g., missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # ANALYTICS ENDPOINTS
  # ============================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained via OAuth 2.0 / OpenID Connect.
        
        **Token Lifetime:** 1 hour
        
        **Required Claims:**
        - `sub` - User ID
        - `scope` - Permission scopes
        - `tenant_id` - Multi-tenant identifier
        
        **Example:**
        ```
        Authorization: Bearer eyJhbGciOiJSUzI1NiIs...
        ```

  # Reusable Parameters
  parameters:
    ArticleIdParam:
      name: articleId
      in: path
      required: true
      description: Unique article identifier
      schema:
        type: string
        pattern: '^art_[a-zA-Z0-9]{8,}$'
      example: "art_abc12345"
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    CategoryParam:
      name: category
      in: query
      description: Filter by category name
      schema:
        type: string

  # Reusable Headers
...
  schemas:
...
    Article:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        title:
          type: string
        content:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, published, archived]
        relevanceScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          readOnly: true
        createdDate:
          type: string
          format: date-time
          readOnly: true
        lastUpdated:
          type: string
          format: date-time
          readOnly: true
        viewCount:
          type: integer
          readOnly: true
        author:
          type: string
          readOnly: true
        version:
          type: integer
          readOnly: true
...
    ArticleCreateRequest:
      type: object
      required:
        - title
        - content
        - category
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 200
        content:
          type: string
          minLength: 50
          maxLength: 50000
        category:
          type: string
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        status:
          type: string
          enum: [draft, published]
          default: draft
    ArticleUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 200
        content:
          type: string
          minLength: 50
          maxLength: 50000
        category:
          type: string
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        status:
          type: string
          enum: [draft, published, archived]
    ArticleListResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/Article'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        hasMore:
          type: boolean

    # Analytics Schemas
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error context
        correlationId:
          type: string
          description: Request ID for support inquiries
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string

  # Reusable Responses
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "BAD_REQUEST"
            message: "Invalid request parameters"
            details:
              invalidParams: ["pageSize must be between 1 and 100"]
            correlationId: "req_xyz789"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "UNAUTHORIZED"
            message: "Authentication required"
            correlationId: "req_xyz789"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "FORBIDDEN"
            message: "Insufficient permissions to access this resource"
            details:
              requiredScope: "articles:write"
            correlationId: "req_xyz789"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "NOT_FOUND"
            message: "Article not found"
            details:
              articleId: "art_abc12345"
            correlationId: "req_xyz789"
    Conflict:
      description: Resource conflict (e.g., concurrent modification)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "CONFLICT"
            message: "Article was modified by another user"
            details:
              currentVersion: 5
              yourVersion: 4
            correlationId: "req_xyz789"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            code: "VALIDATION_ERROR"
            message: "Request validation failed"
            fieldErrors:
              - field: "title"
                message: "Title must be at least 5 characters"
                code: "MIN_LENGTH"
              - field: "content"
                message: "Content is required"
                code: "REQUIRED"
            correlationId: "req_xyz789"
...